Add files via upload
I was able to get both levels working with a mouse interface main menu for choosing the level in both keyboard/mouse mode and VR mode.
The current versions of the files I am submitting are compatible with the VR version of the game.

NEXT:  Another level?  Fix the issue with not being able to shoot zombies if they're too close?